src/app/requirements.txt
Flask
flask_restful
pandas
numpy
gunicorn
tensorflow
keras
sklearn
datetime
scipy
statsmodels
python-dateutil
datetime
xgboost
